reduce library API surface
disables any API that is specific to building a binary. Thus CLI-specific API is not exposed in the library API.

Also reviewed docs
