+++ cpp-linter/README.md
@@ -12,10 +12,6 @@ This crate contains the the library used as a backend for the
   - [x] file annotation
   - [x] pull request review suggestions
 
-Since the [cpp-linter python package][pypi-org] now uses this library
-as a binding, the native binary's `main()` behavior is also present in this
-library (see [`run::run_main()`](fn@crate::run::run_main)).
-
 See also the [CLI document hosted on github][gh-pages].
 
 [pypi-org]: https://pypi.org/project/cpp-linter
