+++ cpp-linter/src/rest_api/mod.rs
@@ -21,11 +21,16 @@ use crate::clang_tools::ClangVersions;
 use crate::cli::{FeedbackInput, LinesChangedOnly};
 use crate::common_fs::{FileFilter, FileObj};
 
+/// The comment marker used to identify bot comments from other comments (from users or other bots).
 pub static COMMENT_MARKER: &str = "<!-- cpp linter action -->\n";
+
+/// The user outreach message displayed in bot comments.
 pub static USER_OUTREACH: &str = concat!(
     "\n\nHave any feedback or feature suggestions? [Share it here.]",
     "(https://github.com/cpp-linter/cpp-linter-action/issues)"
 );
+
+/// The user agent string used for HTTP requests.
 pub static USER_AGENT: &str = concat!("cpp-linter/", env!("CARGO_PKG_VERSION"));
 
 /// A structure to contain the different forms of headers that
